WebMobile/AppMobile/urls.py
from django.urls import path, include
from . import views
from rest_framework import routers

router = routers.DefaultRouter()
router.register('products', views.ProductViewSet)
router.register('categories', views.CategoryViewSet)

urlpatterns = [
    # path('products/<int:product_id>/',)
    path('', include(router.urls)),
]
